Move getIsNavbarOpen to selectors file to get rid of circular deps
