+++ frontend/src/metabase/nav/containers/AppBar/AppBar.tsx
@@ -3,9 +3,9 @@ import { withRouter } from "react-router";
 import _ from "underscore";
 import Collections from "metabase/entities/collections";
 import { logout } from "metabase/auth/actions";
-import { closeNavbar, getIsNavbarOpen, toggleNavbar } from "metabase/redux/app";
-import type { RouterProps } from "metabase/selectors/app";
+import { closeNavbar, toggleNavbar } from "metabase/redux/app";
 import {
+  getIsNavbarOpen,
   getIsCollectionPathVisible,
   getIsLogoVisible,
   getIsNavBarEnabled,
@@ -14,6 +14,8 @@ import {
   getIsQuestionLineageVisible,
   getIsSearchVisible,
 } from "metabase/selectors/app";
+import type { RouterProps } from "metabase/selectors/app";
+
 import { getUser } from "metabase/selectors/user";
 import type { State } from "metabase-types/store";
 import AppBar from "../../components/AppBar";
